refactor app logic into internal/app package
